Tidying up the CSS file, toc class names, en dashes in Translator Notes...
Regexed in K&R-style brackets for the CSS which will save a few KB and look tidier.

Made a new h3.toc class and re-ordered the classes to match the XHTML.

Downloaded warning triangle and pencil icons for the Note and Warning/Caution boxes. I'll need to check 36px rounded black icons are what's best for this.