Renamed fig04-05 diagrams to be shorter and more consistent.
Converted the bullet list below fig4-6 to a numbered list for consistency (I think it looks better) and wrote out the annotations for fig4-6.

I also removed the table for fig4-7 so it should look a lot better now.

I just need to crop the last two images in basic medicine and then the chapter is done.